Deprecate usage of XML Spring schemas
Simplify the code to avoid tons of duplication

Related to #141.
